reorg on objectives. Move to its own module
